requirements.txt
aws-encryption-sdk
boto3
pycryptodome
